+++ docs/releasenotes/releasenotes.rst
@@ -15,6 +15,7 @@ Fixes
 * Fix for network errors resulting from undefined background image
 * Fix: Remove text shadow for blocks without background
 * Dials: remove off level for devices with hidden off level
+* Weather: language for owm3
 
 
 V3.13.0 beta (14-7-2024)